benchmarks/CIFAR_realtime_benchmark/requirements.txt
tensorflow-gpu
numpy
